Move FlagUtils to :common
